refactor: Role model, routes, controller, and frontend
