+++ Models/Role.php
@@ -103,155 +103,6 @@ class Role extends RoleBase
     }
 
     //-------------------------------------------------
-    public static function createItem($request)
-    {
-
-        $inputs = $request->all();
-
-        $validation = self::validation($inputs);
-        if (!$validation['success']) {
-            return $validation;
-        }
-
-
-        // check if name exist
-        $item = self::withTrashed()->where('name', $inputs['name'])->first();
-
-        if ($item) {
-            $response['success'] = false;
-            $response['messages'][] = "This name is already exist.";
-            return $response;
-        }
-
-        // check if slug exist
-        $item = self::withTrashed()->where('slug', $inputs['slug'])->first();
-
-        if ($item) {
-            $response['success'] = false;
-            $response['messages'][] = "This slug is already exist.";
-            return $response;
-        }
-
-        $item = new self();
-        $item->fill($inputs);
-        $item->slug = Str::slug($inputs['slug']);
-        $item->save();
-
-        $response['success'] = true;
-        $response['data']['item'] = $item;
-        $response['messages'][] = 'Saved successfully.';
-        return $response;
-
-    }
-
-    //-------------------------------------------------
-    public function scopeGetSorted($query, $filter)
-    {
-
-        if(!isset($filter['sort']))
-        {
-            return $query->orderBy('id', 'desc');
-        }
-
-        $sort = $filter['sort'];
-
-
-        $direction = Str::contains($sort, ':');
-
-        if(!$direction)
-        {
-            return $query->orderBy($sort, 'asc');
-        }
-
-        $sort = explode(':', $sort);
-
-        return $query->orderBy($sort[0], $sort[1]);
-    }
-    //-------------------------------------------------
-    public function scopeIsActiveFilter($query, $filter)
-    {
-
-        if(!isset($filter['is_active'])
-            || is_null($filter['is_active'])
-            || $filter['is_active'] === 'null'
-        )
-        {
-            return $query;
-        }
-        $is_active = $filter['is_active'];
-
-        if($is_active === 'true' || $is_active === true)
-        {
-            return $query->whereNotNull('is_active');
-        } else{
-            return $query->whereNull('is_active');
-        }
-
-    }
-    //-------------------------------------------------
-    public function scopeTrashedFilter($query, $filter)
-    {
-
-        if(!isset($filter['trashed']))
-        {
-            return $query;
-        }
-        $trashed = $filter['trashed'];
-
-        if($trashed === 'include')
-        {
-            return $query->withTrashed();
-        } else if($trashed === 'only'){
-            return $query->onlyTrashed();
-        }
-
-    }
-    //-------------------------------------------------
-    public function scopeSearchFilter($query, $filter)
-    {
-
-        if(!isset($filter['q']))
-        {
-            return $query;
-        }
-        $search = $filter['q'];
-        $query->where(function ($q) use ($search) {
-            $q->where('name', 'LIKE', '%' . $search . '%')
-                ->orWhere('slug', 'LIKE', '%' . $search . '%');
-        });
-
-    }
-    //-------------------------------------------------
-    public static function getList($request)
-    {
-        $list = self::getSorted($request->filter);
-        $list->isActiveFilter($request->filter);
-        $list->trashedFilter($request->filter);
-        $list->searchFilter($request->filter);
-
-        $rows = config('vaahcms.per_page');
-
-        if($request->has('rows'))
-        {
-            $rows = $request->rows;
-        }
-
-        $list = $list->paginate($rows);
-
-        $countPermissions = Permission::count();
-        $countUsers = User::count();
-
-        $response['success'] = true;
-        $response['data'] = $list;
-        $response['totalPermissions'] = $countPermissions;
-        $response['totalUsers'] = $countUsers;
-
-        return $response;
-
-
-    }
-
-    //-------------------------------------------------
     public static function updateList($request)
     {
 
@@ -300,40 +151,6 @@ class Role extends RoleBase
                 self::whereIn('id', $items_id)->restore();
                 break;
         }
-
-        $response['success'] = true;
-        $response['data'] = true;
-        $response['messages'][] = 'Action was successful.';
-
-        return $response;
-    }
-
-    //-------------------------------------------------
-    public static function deleteList($request): array
-    {
-        $inputs = $request->all();
-
-        $rules = array(
-            'type' => 'required',
-            'items' => 'required',
-        );
-
-        $messages = array(
-            'type.required' => 'Action type is required',
-            'items.required' => 'Select items',
-        );
-
-        $validator = \Validator::make($inputs, $rules, $messages);
-        if ($validator->fails()) {
-
-            $errors = errorsToArray($validator->errors());
-            $response['failed'] = true;
-            $response['messages'] = $errors;
-            return $response;
-        }
-
-        $items_id = collect($inputs['items'])->pluck('id')->toArray();
-        self::whereIn('id', $items_id)->forceDelete();
 
         $response['success'] = true;
         $response['data'] = true;
@@ -706,16 +523,4 @@ class Role extends RoleBase
         return response()->json($response);
     }
     //-------------------------------------------------
-    public static function getModuleSections($request)
-    {
-        $sections = Permission::where('module', $request->module)->withTrashed()->get()->unique('section')->pluck('section');
-
-        $response['success'] = true;
-        $response['data'] = $sections;
-
-        return $response;
-    }
-    //-------------------------------------------------
-
-
 }
